Insure that only specific user attached to bootcamp is allow to update and delete

+++ controllers/bootcamps.js
@@ -46,25 +46,38 @@ exports.createBootCamp = asyncHandler(async (req, res, next) => {
 //@route PUT /api/v1/bootcamps/:id
 //access private
 exports.updateBootCamp = asyncHandler(async (req, res, next) => {
-  const bootcamp = await BootCamp.findByIdAndUpdate(req.params.id, req.body, {
-    new: true,
-    runValidators: true,
-  });
+  let bootcamp = await BootCamp.findById(req.params.id);
   if (!bootcamp) {
     return next(
       new ErrorResponse(`Bootcamp not Found with id of ${req.params.id}`, 404)
     );
   }
+  //make sure user is bootcamp owner
+  if(bootcamp.user.toString() !== req.user.id && req.user.role !=='admin'){
+    return next(
+      new ErrorResponse(`User ${req.params.id} is not authorize to update this bootcamp`, 401)
+    );
+  }
+  bootcamp=await BootCamp.findByIdAndUpdate(req.params.id,req.body,{
+    new:true,
+    runValidators:true
+  })
   res.status(200).json({ success: true, data: bootcamp });
 });
 //@desc delete bootcamp
 //@route DELETE /api/v1/bootcamps/:id
 //access private
 exports.deleteBootCamp = asyncHandler(async (req, res, next) => {
-  const bootcamp = await BootCamp.findById(req.params.id);
+  let bootcamp = await BootCamp.findById(req.params.id);
   if (!bootcamp) {
     return next(
       new ErrorResponse(`Bootcamp not Found with id of ${req.params.id}`, 404)
+    );
+  }
+   //make sure user is bootcamp owner
+  if(bootcamp.user.toString() !== req.user.id && req.user.role !=='admin'){
+    return next(
+      new ErrorResponse(`User ${req.params.id} is not authorize to Delete this bootcamp`, 401)
     );
   }
   await bootcamp.deleteOne();
@@ -105,6 +118,12 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
   if (!bootcamp) {
     return next(
       new ErrorResponse(`Bootcamp not Found with id of ${req.params.id}`, 404)
+    );
+  }
+  //make sure user is bootcamp owner
+  if(bootcamp.user.toString() !== req.user.id && req.user.role !=='admin'){
+    return next(
+      new ErrorResponse(`User ${req.params.id} is not authorize to update this bootcamp`, 401)
     );
   }
   if(!req.files){
